+++ pkg/profiling/task/network/analyze/layer7/events.go
@@ -29,7 +29,7 @@ import (
 )
 
 func (l *Listener) initSocketDataQueue(parallels, queueSize int, config *profiling.TaskConfig) {
-	l.socketDataQueue = btf.NewEventQueue(parallels, queueSize, func(num int) btf.PartitionContext {
+	l.socketDataQueue = btf.NewEventQueue("socket data resolver", parallels, queueSize, func(num int) btf.PartitionContext {
 		return NewSocketDataPartitionContext(l, config)
 	})
 }